+++ e2e/__tests__/babelPluginJestHoist.test.ts
@@ -15,8 +15,8 @@ beforeEach(() => {
   run('yarn', DIR);
 });
 
-it('sucessfully runs the tests inside `babel-plugin-jest-hoist/`', () => {
+it('successfully runs the tests inside `babel-plugin-jest-hoist/`', () => {
   const {json} = runWithJson(DIR, ['--no-cache', '--coverage']);
   expect(json.success).toBe(true);
-  expect(json.numTotalTestSuites).toBe(3);
+  expect(json.numTotalTestSuites).toBe(4);
 });